requirements.txt
aiogram==3.4.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
openai==1.3.7
python-dotenv==1.0.0
asyncio